requirements.txt
streamlit==1.26.0
torch==2.2.0
torchvision
numpy==1.26.2
pillow==9.5.0
scikit-learn==1.3.2
tqdm==4.66.1
pandas==2.2.2
sqlalchemy==2.0.45
opencv-python==4.8.1.78
kaggle==1.5.15
python-multipart==0.0.6